+++ Demo/Keyboard/Demo/DemoKeyboardActionHandler.swift
@@ -24,6 +24,9 @@ class DemoKeyboardActionHandler: StandardKeyboardActionHandler {
     
     private let toastContext: KeyboardToastContext
     
+    
+    // MARK: - Overrides
+    
     override func action(for gesture: KeyboardGesture, on action: KeyboardAction) -> GestureAction? {
         if gesture == .longPress, let action = longPressAction(for: action) { return action }
         if gesture == .tap, let action = tapAction(for: action) { return action }